Give polygon random color upon creation
This change also involves moving the canvas settings from the pen to the polygon. It makes more sense that the pen should be concerned only with drawing what it's provided in its draw method, presumably a polygon with its own canvas settings.

Polygons are also, quite usefully, drawn with a bounding box.