Give polygon translated clones to implement wrapping
This change adds a cloning and translation process to the moving polygon by adding a ring of translated clones around it. This allows the moving polygon to wrap around the canvas bounds immediately when it reaches them. Having a ring of 8 translated clones perhaps accounts for the future possibility in which polygons can shift in direction.